Extract `IfThenCorrector` from `Style/OneLineConditional` to replace an `if-then` structure with a multiline if.
This is necessary in order to avoid rewriting this logic for `Style/IfInsideElse` to be able to rewrite a nested `if-then`.
