+++ lib/rubocop.rb
@@ -139,6 +139,7 @@ require_relative 'rubocop/cop/correctors/condition_corrector'
 require_relative 'rubocop/cop/correctors/each_to_for_corrector'
 require_relative 'rubocop/cop/correctors/empty_line_corrector'
 require_relative 'rubocop/cop/correctors/for_to_each_corrector'
+require_relative 'rubocop/cop/correctors/if_then_corrector'
 require_relative 'rubocop/cop/correctors/lambda_literal_to_method_corrector'
 require_relative 'rubocop/cop/correctors/line_break_corrector'
 require_relative 'rubocop/cop/correctors/multiline_literal_brace_corrector'
