+++ lib/rubocop/cop/style/one_line_conditional.rb
@@ -45,7 +45,7 @@ module RuboCop
 
           message = message(node)
           add_offense(node, message: message) do |corrector|
-            corrector.replace(node, replacement(node))
+            autocorrect(corrector, node)
           end
         end
 
@@ -55,17 +55,22 @@ module RuboCop
           format(MSG, keyword: node.keyword)
         end
 
-        def replacement(node)
+        def autocorrect(corrector, node)
           if always_multiline? || cannot_replace_to_ternary?(node)
-            multiline_replacement(node)
+            IfThenCorrector.new(node, indentation: indentation_width).call(corrector)
           else
-            replaced_node = ternary_replacement(node)
-            return replaced_node unless node.parent
-            return "(#{replaced_node})" if %i[and or].include?(node.parent.type)
-            return "(#{replaced_node})" if node.parent.send_type? && node.parent.operator_method?
-
-            replaced_node
+            corrector.replace(node, ternary_correction(node))
           end
+        end
+
+        def ternary_correction(node)
+          replaced_node = ternary_replacement(node)
+
+          return replaced_node unless node.parent
+          return "(#{replaced_node})" if %i[and or].include?(node.parent.type)
+          return "(#{replaced_node})" if node.parent.send_type? && node.parent.operator_method?
+
+          replaced_node
         end
 
         def always_multiline?
@@ -74,36 +79,6 @@ module RuboCop
 
         def cannot_replace_to_ternary?(node)
           node.elsif_conditional?
-        end
-
-        def multiline_replacement(node, indentation = nil)
-          indentation = ' ' * node.source_range.column if indentation.nil?
-          if_branch_source = node.if_branch&.source || 'nil'
-          elsif_indentation = indentation if node.respond_to?(:elsif?) && node.elsif?
-          if_branch = <<~RUBY
-            #{elsif_indentation}#{node.keyword} #{node.condition.source}
-            #{indentation}#{branch_body_indentation}#{if_branch_source}
-          RUBY
-          else_branch = else_branch_to_multiline(node.else_branch, indentation)
-          if_branch + else_branch
-        end
-
-        def else_branch_to_multiline(else_branch, indentation)
-          if else_branch.nil?
-            'end'
-          elsif else_branch.if_type? && else_branch.elsif?
-            multiline_replacement(else_branch, indentation)
-          else
-            <<~RUBY.chomp
-              #{indentation}else
-              #{indentation}#{branch_body_indentation}#{else_branch.source}
-              #{indentation}end
-            RUBY
-          end
-        end
-
-        def branch_body_indentation
-          ' ' * (@config.for_cop('Layout/IndentationWidth')['Width'] || 2)
         end
 
         def ternary_replacement(node)
@@ -140,6 +115,10 @@ module RuboCop
           return true if node.respond_to?(:prefix_not?) && node.prefix_not?
 
           node.respond_to?(:arguments?) && node.arguments? && !node.parenthesized_call?
+        end
+
+        def indentation_width
+          @config.for_cop('Layout/IndentationWidth')['Width']
         end
       end
     end
